requirements/prod.txt
cython==0.29.23
pafy==0.5.5
scikit-learn==0.24.2
confluent-kafka==1.7.0
kafka-python==2.0.2
numpy==1.19.5
clickhouse-driver==0.2.1
ffmpeg_python==0.2.0
streamlink==2.1.2
audioread==2.1.9
librosa==0.8.1
wavio==0.0.4
kapre==0.3.5
matplotlib==3.3.4
scipy==1.5.4
pandas==1.1.5
tqdm==4.61.0